+++ backend_new/src/routes/intraSports/repository.ts
@@ -107,7 +107,7 @@ export async function updateActivity(activityData: IntraSports, id: string) {
     }
 }
 
-export async function deleteActivity(id: string) {
+export async function deleteActivity(id: string, role: string, accessTo: string) {
     try {
         const getData = await db.query.intraSports.findFirst({
             where: eq(intraSports.id, id),
@@ -115,8 +115,11 @@ export async function deleteActivity(id: string) {
         if (!getData) {
             return { status: 404, message: 'Not Found' };
         }
-        await db.delete(intraSports).where(eq(intraSports.id, id));
-        return { message: 'Delete successful' };
+        if(role === 'admin' && (accessTo === 'all' || accessTo === 'student')){
+            await db.delete(intraSports).where(eq(intraSports.id, id));
+            return { message: 'Delete successful' };
+        }
+        return { status: 403, message: 'Forbidden' };
     } catch (error) {
         console.log(error);
         errs(error);
