+++ backend_new/src/routes/studentCareerCounselling/repository.ts
@@ -97,7 +97,7 @@ export async function updateActivity(activityData: StudentCareerCounselling, id:
     }
 }
 
-export async function deleteActivity(id: string) {
+export async function deleteActivity(id: string,role: string, accessTo: string) {
     try {
         const getData = await db.query.studentCareerCounselling.findFirst({
             where: eq(studentCareerCounselling.id, id),
@@ -105,8 +105,11 @@ export async function deleteActivity(id: string) {
         if (!getData) {
             return { status: 404, message: 'Not Found' };
         }
-        await db.delete(studentCareerCounselling).where(eq(studentCareerCounselling.id, id));
-        return { message: 'Delete successful' };
+        if(role === 'admin' && (accessTo === 'all' || accessTo === 'student')){
+            await db.delete(studentCareerCounselling).where(eq(studentCareerCounselling.id, id));
+            return { message: 'Delete successful' };
+        }
+        return { status: 403, message: 'Forbidden' };
     } catch (error) {
         console.log(error);
         errs(error);
