+++ backend_new/src/routes/studentEntranceExam/repository.ts
@@ -142,7 +142,7 @@ export async function updateActivity(activityData: StudentEntranceExam, id: stri
     }
 }
 
-export async function deleteActivity(id: string) {
+export async function deleteActivity(id: string,role: string, accessTo: string) {
     try {
         const getData = await db.query.studentEntranceExam.findFirst({
             where: eq(studentEntranceExam.id, id),
@@ -150,8 +150,11 @@ export async function deleteActivity(id: string) {
         if (!getData) {
             return { status: 404, message: 'Not Found' };
         }
-        await db.delete(studentEntranceExam).where(eq(studentEntranceExam.id, id));
-        return { message: 'Delete successful' };
+        if (role === 'admin' && (accessTo === 'all' || accessTo === 'student')) {
+            await db.delete(studentEntranceExam).where(eq(studentEntranceExam.id, id));
+            return { message: 'Delete successful' };
+        }
+        return { status: 403, message: 'Forbidden' };
     } catch (error) {
         console.log(error);
         errs(error);
