+++ backend_new/src/routes/studentHigherStudies/repository.ts
@@ -121,7 +121,7 @@ export async function updateActivity(activityData: StudentHigherStudies, id: str
     }
 }
 
-export async function deleteActivity(id: string) {
+export async function deleteActivity(id: string,role:string, accessTo: string) {
     try {
         const getData = await db.query.studentHigherStudies.findFirst({
             where: eq(studentHigherStudies.id, id),
@@ -129,8 +129,11 @@ export async function deleteActivity(id: string) {
         if (!getData) {
             return { status: 404, message: 'Not Found' };
         }
-        await db.delete(studentHigherStudies).where(eq(studentHigherStudies.id, id));
-        return { message: 'Delete successful' };
+        if(role === 'admin' && (accessTo === 'all' || accessTo === 'student')){
+            await db.delete(studentHigherStudies).where(eq(studentHigherStudies.id, id));
+            return { message: 'Delete successful' };
+        }
+        return { status: 403, message: 'Forbidden' };
     } catch (error) {
         console.log(error);
         errs(error);
